pyproject.toml
[project]
name = "llg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.3.0",
    "isort>=8.0.1",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.0.10",
    "python-dotenv>=1.2.2",
]
